+++ pytorch_segmentation_models_trainer/custom_callbacks/__init__.py
@@ -1,11 +1,12 @@
 from .image_callbacks import EnhancedImageSegmentationResultCallback
 from .kl_annealing_callback import KLAnnealingCallback
 from .metrics_callbacks import ConfusionMatrixCallback
-from .training_callbacks import EMACallback
+from .training_callbacks import EMACallback, FinalMetricsCallback
 
 __all__ = [
     "ConfusionMatrixCallback",
     "EnhancedImageSegmentationResultCallback",
     "EMACallback",
+    "FinalMetricsCallback",
     "KLAnnealingCallback",
 ]